+++ public/js/sa_bv.js
@@ -28,3 +28,17 @@ function showSuccess_redirect(message, url){
         document.location.href = url;
     });
 }
+
+// IDR Currency
+function idr_curr(nStr)
+{
+	nStr += '';
+	x = nStr.split('.');
+	x1 = x[0];
+	x2 = x.length > 1 ? '.' + x[1] : '';
+	var rgx = /(\d+)(\d{3})/;
+	while (rgx.test(x1)) {
+		x1 = x1.replace(rgx, '$1' + '.' + '$2');
+	}
+	return "Rp "+x1 + x2;
+}