+++ routes/web.php
@@ -25,7 +25,16 @@ Route::group(['middleware' => ['auth','web',]], function(){
         return view('staff.dashboard.index');
     })->name('staff');
 
-    Route::resource('/staff/kategori', 'KategoriController'); //Kategori
+    //Kategori
+    Route::resource('/staff/kategori', 'KategoriController');
+    Route::get('/data/kategori/{id}', 'KategoriController@kategoriSpecificJson'); //Data Kategori
     Route::get('/list/kategori', 'KategoriController@kategoriAllJson'); //List Kategori
+
+    //Barang
+    Route::resource('/staff/barang', 'BarangController');
+    Route::get('/list/barang', 'BarangController@barangJson'); //List Barang (All)
+    Route::get('/data/barang/{id}', 'BarangController@barangSpecificJson'); //Data Barang
+    Route::get('/data/barang/kategori/{id}', 'BarangController@kategoriSpecificJson'); //Data Barang
+
 });